fix(desktop): fix the type of cotoami.tauri.Event.id

+++ cotoami_desktop/ui/src/main/scala/cotoami/tauri/Event.scala
@@ -57,7 +57,7 @@ trait Event[T] extends js.Object {
   val event: String = js.native
 
   // Event identifier used to unlisten
-  val id: Int = js.native
+  val id: Double = js.native
 
   // Event payload
   val payload: T = js.native
